requirements.txt
ultralytics==8.3.0
torch>=2.0.0
fastapi
uvicorn
opencv-python
streamlit
requests
Pillow
python-multipart
ipykernel
opencv-python
